+++ sources/VS/Editor/src/Form.cpp
@@ -304,7 +304,7 @@ bool Form::ExistPoint(int mouseX, int mouseY, bool pressed)
 
     for(uint i = 0; i < points.size(); i++)
     {
-        double distance = points[i].DistanceFromMouse(Round<int>(mouseX / scaleX), Round<int>(mouseY / scaleY));
+        double distance = points[i].DistanceFromMouse(Round<int>(static_cast<float>(mouseX) / scaleX), Round<int>(static_cast<float>(mouseY) / scaleY));
 
         if(distance < nearestDistance)
         {
@@ -358,11 +358,11 @@ static void DrawForm(const uint16 data[Point::NUM_POINTS], Color color)
 
     for (int i = 1; i < Point::NUM_POINTS; i++)
     {
-        int x0 = Round<int>(scaleX * (i - 1));
-        int y0 = Round<int>(scaleY * (Point::MAX - data[i - 1]));
+        int x0 = Round<int>(scaleX * static_cast<float>(i - 1));
+        int y0 = Round<int>(scaleY * static_cast<float>(Point::MAX - data[i - 1]));
 
-        int x1 = Round<int>(scaleX * i);
-        int y1 = Round<int>(scaleY * (Point::MAX - data[i]));
+        int x1 = Round<int>(scaleX * static_cast<float>(i));
+        int y1 = Round<int>(scaleY * static_cast<float>(Point::MAX - data[i]));
 
         TheCanvas->DrawLine(x0, y0, x1, y1);
     }
@@ -380,15 +380,15 @@ void Form::Draw()
 
     for (Point point : points)
     {
-        int x = Round<int>(scaleX * point.pos);
-        int y = Round<int>(scaleY * (Point::MAX - point.data));
+        int x = Round<int>(scaleX * static_cast<float>(point.pos));
+        int y = Round<int>(scaleY * static_cast<float>(Point::MAX - point.data));
 
         TheCanvas->SetPoint(x, y, Point::SIZE);
     }
 
     if (iCurPoint != static_cast<uint>(-1))
     {
-        TheCanvas->SetPoint(Round<int>(scaleX * points[iCurPoint].pos), Round<int>(scaleY * (Point::MAX - points[iCurPoint].data)), Point::SIZE * 3);
+        TheCanvas->SetPoint(Round<int>(scaleX * static_cast<float>(points[iCurPoint].pos)), Round<int>(scaleY * static_cast<float>(Point::MAX - points[iCurPoint].data)), Point::SIZE * 3);
     }
 
     if (drawAdditionData)
@@ -403,9 +403,9 @@ void Form::LinearInterpolation(uint16 pos1, uint16 pos2)
     uint16 data1 = data[pos1];
     uint16 data2 = data[pos2];
 
-    float delta = static_cast<float>(data2 - data1) / (pos2 - pos1);  // Разность значений между соседними точками
+    float delta = static_cast<float>(data2 - data1) / static_cast<float>(pos2 - pos1);  // Разность значений между соседними точками
     
-    float value = data[pos1] + delta;                                  // Значение в текущей позиции
+    float value = static_cast<float>(data[pos1]) + delta;                                  // Значение в текущей позиции
     
     for (int i = pos1 + 1; i < pos2; i++)
     {
